third_party/WebKit/LayoutTests/fast/events/platform-wheelevent-paging-y-in-non-scrolling-page-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS event.wheelDeltaY is window.givenScrollTop * -3
PASS event.wheelDeltaX is 0
PASS event.wheelDelta is window.givenScrollTop * -3
PASS document.scrollingElement.scrollTop is window.expectedScrollTop
PASS document.scrollingElement.scrollLeft is window.expectedScrollLeft
